AB#4308 samples table & project dashboard - time is human readable
Changed the samples date created and updated columns to not show utc time but local time its not as custom as the dashboards but significantly more readable in a table format. The project dashboard time was changed in the same way as the home dashboard that can be seen in the previous commit.