AB#4627 Identified the problem and made a bandaid solution will need a better one.
This just checks if the column headers are 0 and if they are it needs to set the loading of the samples table to be false. However when this happens it needs to do gracefully without any warning which is what is happening at the moment. Will need to discuss how to properly fix this issue.