AB#4716 More visual changes to the proforma page
This change encapsulates style changes and some design changes so that the field snapshot fits nicely on the card media.

The change was made to make a horizontal stack however due mui components having different properties in the space they take this was reverted back to a grid.

Expand button/info text changed

The info icon is also better placed.

Otherwise refactoring was done to make the data flow and component generation more readable.