AB#4650 enhanced the detailed sample page to have better error handling and route generation
when a user types out a custom path for a sample and a group and they don't have
access to the group or the group does not exist it will show a snack bar saying
that their query is incorrect.

for the path generation I used the same method from the tree page so when the
selected group is changed by the user or automatically chosen in some edge
cases it is put in the path thus can be shared which will change the state of
the page of the user that receives the link.