New tab that holds a table of the databases with management options
this table give the user the option to delete tables with a confirmation if they would like

simple implementation which can be evolved further at a later stage.

Some ideas could be to group the databases by user, fields etc

Also if this chrono view is shown we need to think of better ways to display the list of fields on a table

doesn't have to be a table.