AB#5273 userSlice with redux, permission system, role based components in dataset table
userslice will load the user roles and info for now and is loaded in the app.tsx. Thus some restructuring has happened.

the table has new columns for actions and the created by.

The table will change the interaction with the table actions based on roles

basic role table included with a function to determine perms on ui elements