AB#5244 Breadcrumbs removed if leaf
Main issue was since use navigate was not being used the breadcrumbs on the project page would not update depending on the tab selection of the user.

Could not update underlying state of the url without creating an events to listen to the url state, which would lead us to making the same state system that react is supposed to do for us. Adding to this creating a custom listener of the url is not supported by every browser as it has been made obsolete by react in some places.

Thus on the project page the leaf nodes that reference the tab will no longer appear as it was a little bit redundant when the user could navigate with the tab buttons anyway. However if the user is deeper e.g. the Plots page they will be able to see the plots breadcrumb which will link them to the appropriate tab to return to.