AB#5267 Has_sequences is no longer null
Both the metadata view load and the direct query to the database by the organisation page both have the Has_sequences field altered to make null false. This is done at the use effect for the org sample page and for the project sample page this alteration is handled by the slice.

The org page fix is temporary as that page will also work of views in the future.