AB#4811 User Info Is fetched immediately, new routes, new navigation button
The perms are fetched upfront because I want routes/urls that get passed around to only be active if the permission allows it.

In this case most of these urls that are going to be added for admin management are restricted views only admins can access. Thus, I believe the urls to these pages should not exist if a user does not have admin privileges.