AB#5451 can exclude and search for empty/null values
refactored some code to be readable.

registered a generic null filter thing for each field where the value is the row data and the filter is the actual value you want filtered on. I made the filter value a boolean so if its true it shows the null row and if its false it excludes it This was done due to the fact that each field can only have one stand alone custom field.

In the data preprocessing I made it so that if a date string is empty it gets replaced by a null as the date filtering falls over if it encounters a empty string type.