AB#5544 AB#5546 AB#5547 Chips removable/Patch possible
Chips when removed also trigger a re-render so that the when a user clicks the remove icon there is visual feedback in what should happen

states are kept for the editable version of the user and the stable/real version so if a user cancels the edit operation real state is restored.

Patch requests with body is sent and works with minor edits

now need to think about drop downs for fields which cannot be entirely custom strings. Need to think about drop Downs for org changes and what the effects of that are might be annoying