AB#5658 reroot state variable which controls root
can now reroot from midpoint which then leads to the button to change to reset.
when midpoint is active the tree stays within that format however not sure if the jump to subtree is working as intended with this feature
the title of the tree reflects the state of the rooting.