AB#5815 fixed the plots not showing all data on first load.
The filtered data was not getting initialised with the base set on first load and it was relying on the Data table to initialise the filtered data set on first render. This behaviour was not correct.

Thus the metadata will be passed to filtered data as soon it has loaded from redux.