AB#5938 removed the use of input filter on the main tables
Tables for the Org, Proj, and Tree pages are no longer supporting the inputfilter/filterlist object for filtering the table it's all DataTableFilterMeta objects that directly work with PrimeReact.

Next work is to change the way the widgets get to the samples page and apply their filters.

chips no longer shake when a duplicate gets added. I think it's not needed and it keeps the objects cleaner. I say this, as when a user tries to add a duplicate filter it already doesn't apply and a error toast pop comes up.