AB#6246 fixed the state issue have to introduce a useEffect
The initial value need to be set if it differs from the state, this was done to keep consistency with all the stats at play
Url navigation, Url data accuracy and table and filter accuracy.

Maybe its functionality is too overloaded wouldn't know at the moment what an alternative approach would be.