AB#6613 added date_coll aggregation with better colours and layout
Extra widget included mimics accessions count implementation, colours adjusted to work with vega and the env settings
all bars clickable.

layout was changed as I changed the Grid component to the newer version from mui which required some refactoring and simplification.

Mistake Noticed that needs to be fixed for next time is the filters on click are generic and not bound to the org. A small oversight due to the test setup limitations for the moment