AB#6685 Added dialogue to confirm user changes.
on selection change or it should be named on selection add the display structure which is a copy of the original is changed to show the changes in real time
and a post payload is created for the privilege assignment.

User can then press the save icon to show a dialogue with a list of the additions that you will add with a final confirm or cancel.