AB#6953 Large amounts of legends would break plot - Fixed
now, columns aren't decided dynamically. they're based on a spec defined in utils.

we could still make it dynamic if needed. I think the main issue was the `orient` property being set to `right` instead of `bottom`.

we'll need to test in prod to be sure. I was sort of able to recreate the error locally by adding a bunch of unique `st` values, but there's a chance this is a completely different issue that I'm solving.