AB#7158 initial commit with functional apache charts implementation
New packages and tsconfig changes to support apache echarts.

geo json of aus and nz

MapTest renders the chart with the appropriate options
this also holds the filter and the colour selection component

MapTest detail dispatches the redux slice to be ready for the chart,
and also passes the map.