AB#7158 Map Detail handles all the control related components
This should also handle the filter layer/processing for the maps rendering component.
Decided for this so the controls are sepereate from the rendering logic as both are quite complex.

did not move aggregation logic here as that is closely tied to how the map is rendered. Could be argued the other way as-well.