Removed duplicate in geojson and split useEffects to seperate concerns
The useEffect needed to be split because there was a race condition between determining the ISO code for indexing and loading the actual GeoJSON.
If the ISO code index is iso_3 but the map JSON is still in region mode, duplicates occur, causing a crash: every region’s iso_3 code matches its parent country, so multiple regions produce duplicate hits.

This change trys to ensure ISO code indexing and GeoJSON processing happen in the correct order, preventing duplicates and crashes.