AB#7452 implemented the conditional rendering of the share button
The Share button on the main organisation page is conditionally rendered based on v1 roles. It is shown if the user is an AusTrakkaAdmin; otherwise, it is shown only if the user has the Uploader role for the organisation they are viewing.

This may need another pass to get the edge-cases right.